+++ Glossary/ConosleColor.cs
@@ -1,0 +1,6 @@
+﻿namespace Glossary
+{
+    public class ConosleColor
+    {
+    }
+}
