+++ Glossary/Service.cs
@@ -8,178 +8,47 @@ namespace Glossary
 {
     public class Service
     {
-        readonly List<ConsoleColor> _colors = new();
-        readonly List<Space> _spaces = new();
-        private Pixel[,]? _pixels;
-        private int _width;
-        private int _height;
-
-        /// <summary>
-        /// Сообщение об ошибке программы
-        /// </summary>
-        /// <param name="ex"></param>
-        private void WriteExMsg(string msg)
-        {
-            Console.ForegroundColor = ConsoleColor.Red;
-            Console.WriteLine(msg);
-            Console.ForegroundColor = ConsoleColor.White;
-        }
-
-
-        /// <summary>
-        /// Выбор размера поля для игры
-        /// </summary>
-        private void SizeSelection()
-        {
-            while (true)
-            {
-                try
-                {
-                    Console.Write("Введите число (ширину окна) от 2 до 80: ");
-                    _width = int.Parse(Console.ReadLine());
-                        
-                    Console.Write("Введите число (высоту окна) от 2 до 50: ");
-                    _height = int.Parse(Console.ReadLine());
-                    if (_width >=2 && _height >=2 && _width<=80&&_height<=50)
-                       break;
-                    Console.WriteLine("Введите значения в нужном диапазоне!");
-                }
-                catch (Exception ex)
-                {
-                    WriteExMsg(ex.Message);
-                }
-            }
-        }
-
-
-
-        /// <summary>
-        /// Выбор цвета для игры
-        /// </summary>
-        private void ColorSelection()
-        {
-            while (true)
-            {
-                try
-                {
-                    List<ConsoleColor> colorList = new();
-                    int i = 0;
-                    foreach (var item in Enum.GetValues<ConsoleColor>())
-                    {
-                        if (item == ConsoleColor.Black)
-                            continue;
-                        Console.WriteLine($"{++i}) {item}");
-                        colorList.Add(item);
-                    }
-                    Console.Write("Выберите цвета которые будут участвовать в игре, напишите цифры через пробел: ");
-                    string[] numbers = Console.ReadLine().Split(' ');
-
-                    foreach (var item in numbers)
-                    {
-                        _colors.Add(colorList[int.Parse(item) - 1]);
-                    }
-                    break;
-                }
-                catch (Exception ex)
-                {
-                    WriteExMsg(ex.Message);
-                }
-            }
-
-        }
-
-        /// <summary>
-        /// Заполнет поле случайными цветами
-        /// </summary>
-        private void Fill()
-        {
-            Console.Clear();
-            _pixels = new Pixel[_width, _height];
-            for (int y = 0; y < _width; y++)
-            {
-                for (int x = 0; x < _height; x++)
-                {
-                    ConsoleColor color = _colors[Random.Shared.Next(_colors.Count)];      //Выбор рандомного цвета
-                    Console.BackgroundColor = color;
-                    Console.Write(' ');
-                    _pixels[x, y] = new Pixel(x, y, color);     // Добавление данного пикселя в список
-                }
-                Console.BackgroundColor = ConsoleColor.Black;
-                Console.WriteLine();
-            }
-            Console.SetCursorPosition(0, 0);
-        }
-
-        /// <summary>
-        /// Находит все пространства
-        /// </summary>
-        private void SpaceCreator()
-        {
-            int countOfSpaces = 0;
-            for (int y = 0; y < _width; y++)
-            {
-                for (int x = 0; x < _height; x++)
-                {
-                    if (_pixels[x, y].IsAdded)
-                        continue;
-
-                    _spaces.Add(new Space(_pixels, _width, _height));
-                    _spaces[countOfSpaces].CreateSpace(x, y);
-                    countOfSpaces++;
-
-                }
-            }
-
-        }
-
-        private void OneStep()
-        {
-            foreach (var item in _spaces)
-            {
-
-                _pixels = item.ColorARandomAdjacentPixel(_pixels);
-
-            }
-        }
-
-        private void RestartField()
-        {
-            foreach (Pixel? item in _pixels)
-            {
-                item.IsAdded = false;
-            }
-            _spaces.Clear();
-        }
-
-        private void CheckingTheSpaceForEmptiness()
-        {
-            for (int i = 0; i < _spaces.Count; i++)
-            {
-                if (_spaces[i].CheckCountOfPixelInThisSpace())
-                    _spaces.RemoveAt(i);
-            }
-        }
+        private Frontend _front;
+        private Backend _backend;
 
         public void StartGame()
         {
-            SizeSelection();
-            ColorSelection();
-            Fill();
-            Console.CursorVisible = false;
+            string answer;
             while (true)
             {
-                SpaceCreator();
-                OneStep();
-                CheckingTheSpaceForEmptiness();
-                //Thread.Sleep(1);
-                //Console.ReadKey();
-                RestartField();
-                if (_spaces.Count == 1)
+                _front = new();
+                _front.Start();
+                _backend = new(_front.Width,_front.Height, _front.Colors);
+                _backend.Start();
+                
+                if (PlayAgain())
                     break;
             }
 
         }
 
+        private bool PlayAgain()
+        {
+            string answer;
+            while (true)
+            {
+                Console.WriteLine("Хотите начать заново? Напишите 'y' - да, 'n' - нет");
+                answer = Console.ReadLine();
 
+                switch (answer)
+                {
+                    case "n":
+                        return true;
+                    case "y":
+                        return false;
+
+                    default:
+                        Console.ForegroundColor = ConsoleColor.Red;
+                        Console.WriteLine("Введите одну из указанных букв!");
+                        Console.ForegroundColor = ConsoleColor.White;
+                        break;
+                }
+            }
+        }
     }
 }
